+++ CHANGELOG.md
@@ -4,11 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] 
+## [Unreleased] -
 ### Added
 - Added query data to get Agent host information.
 - HAProxy plugin.
 - Systemctl management API call.
+- Database migration for the DHCP service management.
+- In the firewalls/clusters tree new sort of nodes under a firewall or cluster node. 
+- Migrations are added to create the System nodes and their children: DHCP, Keepalived, and HAProxy.
+
+### Changed
+- Optimize and reduce the time for the firewalls and clusters tree creation.
+
+
 
 ## [1.9.2] - 2023-06-07
 ### Added